+++ src/main/java/mods/railcraft/client/render/RenderIronTank.java
@@ -171,7 +171,7 @@ public class RenderIronTank extends TileEntitySpecialRenderer {
                 GL11.glPushMatrix();
 
                 float cap = tank.getCapacity();
-                float level = (float) Math.min(tank.renderData.amount, cap) / cap;
+                float level = Math.min(tank.renderData.amount, cap) / cap;
 
                 bindTexture(FluidRenderer.getFluidSheet(tank.renderData.fluid));
                 FluidRenderer.setColorForTank(tank);
