+++ src/main/java/mods/railcraft/common/blocks/machine/beta/TileBoilerFirebox.java
@@ -102,7 +102,7 @@ public abstract class TileBoilerFirebox extends TileBoiler implements IInventory
             updateLighting();
             if (isBurning() && random.nextInt(100) < 20) {
                 float f = (float) xCoord + 0.5F;
-                float f1 = (float) yCoord + 0.4375F + (random.nextFloat() * 3F / 16F);
+                float f1 = yCoord + 0.4375F + (random.nextFloat() * 3F / 16F);
                 float f2 = (float) zCoord + 0.5F;
                 float f3 = 0.52F;
                 float f4 = random.nextFloat() * 0.6F - 0.3F;
